% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerTest.R
\name{analyse.primers}
\alias{analyse.primers}
\title{Plot Curves For Testing qPCR Primers}
\usage{
analyse.primers(scheme, dil = NULL, notPas = NULL)
}
\arguments{
\item{scheme}{A pipetting scheeme, e.g., generated by get.pipettingScheme.
Generally a data.frame containing gene, cond, col and row information.}

\item{dil}{Conditions and corresponding dilution steps. A named vector with
conditions as names and denominator of dilution steps as values.}

\item{notPas}{Character vector containing wells, that are not to be included
in the analysis.}
}
\description{
Plot Curves For Testing qPCR Primers
}
\examples{
scheme <- get.pipettingScheme(paste("primer", 1:7), 1:5, 1, 2, 1, nrow = 10, first_col = 18)
    scheme <- import.LCcq("data/180927_primer.txt", scheme)
    dils <- c(1, 6, 36, 216, 1298)
    names(dils) <- 1:5
    analyse.primers(scheme, dil = dils)

}
