% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipettingScheme.R
\name{get.pipettingScheme}
\alias{get.pipettingScheme}
\title{Get A Tibble With A Pipetting Scheme For qPCR}
\usage{
get.pipettingScheme(names_gene, names_cond, rep_qpcr, rep_tech,
  rep_biol = 1, first_col = 1, nrow = 16)
}
\arguments{
\item{names_gene}{character vector containing names of the genes measured}

\item{names_cond}{character vector containing names of the conditions}

\item{rep_qpcr}{number of technical replicates in the qPCR per sample}

\item{rep_tech}{number of technical replicates e.g., in cell culture per condition of a biological replicate}

\item{rep_biol}{number of biological replicates}

\item{first_col}{first column with samples}

\item{nrow}{number of rows with samples}
}
\description{
Get A Tibble With A Pipetting Scheme For qPCR
}
\examples{
get.pipettingScheme(names_gene = c("Ucp1", "Gapdh"),
    names_cond = c("NTC", "CL", "iso"), rep_qpcr = 2,
    rep_tech = c(3,3,2), rep_biol = 2)

}
