% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.LCcq}
\alias{import.LCcq}
\title{Import Cq Values Calculated Using LightCycler Software}
\usage{
import.LCcq(file, pipettingScheme = NULL, decimal_mark = ".")
}
\arguments{
\item{file}{A txt file with calculated Cq values from the LightCycler software}

\item{scheme}{A pipetting scheeme, e.g., generated by get.pipettingScheme.
Generally a data.frame containing gene, cond, col and row information.}
}
\description{
Import Cq Values Calculated Using LightCycler Software
}
\examples{
scheme <- get.pipettingScheme(paste("primer", 1:7), c(1:5), 1, 2, 1, nrow = 10, first_col = 18)
    import.LCcq("data-raw/primer.txt", scheme, decimal_mark = ',')

}
